Cast length to size_t in vsnprintf call

+++ pntr.h
@@ -4256,7 +4256,7 @@ PNTR_API void pntr_draw_text_ex(pntr_image* dst, pntr_font* font, int posX, int 
 
     va_list arg_ptr;
     va_start(arg_ptr, text);
-    vsnprintf(output, length, text, arg_ptr);
+    vsnprintf(output, (size_t)length, text, arg_ptr);
     va_end(arg_ptr);
 
     pntr_draw_text(dst, font, output, posX, posY, tint);
